Rename .util.html to .util.parse_html
PyCharm's debugger tries to load our own ".util.html" module instead of
BeautifulSoup's one, quite possibly because we add a .pch file to
virtualenv's site-packages to make Inline::Python work.
