+++ lib/MediaWords/Feed/Scrape.pm
@@ -9,7 +9,7 @@ use Modern::Perl "2015";
 use MediaWords::CommonLibs;
 
 use MediaWords::Feed::Parse;
-use MediaWords::Util::HTML;
+use MediaWords::Util::ParseHTML;
 use MediaWords::Util::URL;
 use MediaWords::Util::Web;
 
@@ -653,7 +653,7 @@ sub _immediate_redirection_url_for_medium($)
     }
 
     my $html = $response->decoded_content || '';
-    $new_url = MediaWords::Util::HTML::meta_refresh_url_from_html( $html );
+    $new_url = MediaWords::Util::ParseHTML::meta_refresh_url_from_html( $html );
     if ( $new_url and ( !MediaWords::Util::URL::urls_are_equal( $medium->{ url }, $new_url ) ) )
     {
         DEBUG "New medium URL via HTML <meta/> refresh: $medium->{url} => $new_url";
