+++ lib/MediaWords/Util/ParseHTML.pm
@@ -1,4 +1,4 @@
-package MediaWords::Util::HTML;
+package MediaWords::Util::ParseHTML;
 
 use strict;
 use warnings;
@@ -6,6 +6,6 @@ use warnings;
 use Modern::Perl "2015";
 use MediaWords::CommonLibs;    # set PYTHONPATH too
 
-import_python_module( __PACKAGE__, 'mediawords.util.html' );
+import_python_module( __PACKAGE__, 'mediawords.util.parse_html' );
 
 1;
