+++ mediacloud/mediawords/util/url/variants.py
@@ -2,7 +2,7 @@ import re
 from typing import List
 
 from mediawords.db import DatabaseHandler
-from mediawords.util.html import link_canonical_url_from_html
+from mediawords.util.parse_html import link_canonical_url_from_html
 from mediawords.util.log import create_logger
 from mediawords.util.perl import decode_object_from_bytes_if_needed
 from mediawords.util.url import fix_common_url_mistakes, is_http_url, normalize_url, is_homepage_url
